+++ src/isapid.cpp
@@ -171,7 +171,7 @@ int ibis::sapid::write(const char* dt) const {
     char header[] = "#IBIS\14\0\0";
     header[5] = (char)ibis::index::SAPID;
     header[6] = (char)(useoffset64 ? 8 : 4);
-    int32_t ierr = UnixWrite(fdes, header, 8);
+    off_t ierr = UnixWrite(fdes, header, 8);
     if (ierr < 8) {
 	LOGGER(ibis::gVerbose > 0)
 	    << "Warning -- " << evt
