+++ src/ixambit.cpp
@@ -982,7 +982,7 @@ int ibis::ambit::write(const char* dt) const {
     char header[] = "#IBIS\2\0\0";
     header[5] = (char)ibis::index::AMBIT;
     header[6] = (char)(useoffset64 ? 8 : 4);
-    int32_t ierr = UnixWrite(fdes, header, 8);
+    off_t ierr = UnixWrite(fdes, header, 8);
     if (ierr < 8) {
 	LOGGER(ibis::gVerbose > 0)
 	    << evt << " failed to write the 8-byte header, ierr = " << ierr;
